Added the ability to load via `require`
Prior to this change, we had to use script tags to load
playground.js. Now, we can simply download playground using
npm, and use CommonJS' `require` to load the module.

This commit adds a `package.json` file for the sake of compatibility
with npm, a `.gitignore` file to ignore the `node_modules` folder,
enhances `plugins/playground.three.js` to be loadable via `require`
and enhances build.js to distribute a CommonJS-equivalent of the
build files.